chore: `events` module restructuring
