+++ src/entities/events/channel/group_join.rs
@@ -1,13 +1,13 @@
-use super::ServerToClientEvent;
+use crate::entities::ServerToClientEvent;
 
-/// A user has started typing in a channel.
+/// A user has joined the group.
 #[derive(Debug)]
-pub struct ChannelStartTypingEvent {
+pub struct ChannelGroupJoinEvent {
     id: String,
     user_id: String,
 }
 
-impl ChannelStartTypingEvent {
+impl ChannelGroupJoinEvent {
     /// Returns the channel id.
     pub fn id(&self) -> &str {
         &self.id
@@ -19,9 +19,9 @@ impl ChannelStartTypingEvent {
     }
 }
 
-impl From<ServerToClientEvent> for ChannelStartTypingEvent {
+impl From<ServerToClientEvent> for ChannelGroupJoinEvent {
     fn from(event: ServerToClientEvent) -> Self {
-        if let ServerToClientEvent::ChannelStartTyping { id, user } = event {
+        if let ServerToClientEvent::ChannelGroupJoin { id, user } = event {
             Self { id, user_id: user }
         } else {
             panic!("An incorrect event was provided: {:?}", event);
