+++ src/entities/events/channel/group_leave.rs
@@ -1,13 +1,13 @@
-use super::ServerToClientEvent;
+use crate::entities::ServerToClientEvent;
 
-/// A user has joined the group.
+/// A user has left the group.
 #[derive(Debug)]
-pub struct ChannelGroupJoinEvent {
+pub struct ChannelGroupLeaveEvent {
     id: String,
     user_id: String,
 }
 
-impl ChannelGroupJoinEvent {
+impl ChannelGroupLeaveEvent {
     /// Returns the channel id.
     pub fn id(&self) -> &str {
         &self.id
@@ -19,9 +19,9 @@ impl ChannelGroupJoinEvent {
     }
 }
 
-impl From<ServerToClientEvent> for ChannelGroupJoinEvent {
+impl From<ServerToClientEvent> for ChannelGroupLeaveEvent {
     fn from(event: ServerToClientEvent) -> Self {
-        if let ServerToClientEvent::ChannelGroupJoin { id, user } = event {
+        if let ServerToClientEvent::ChannelGroupLeave { id, user } = event {
             Self { id, user_id: user }
         } else {
             panic!("An incorrect event was provided: {:?}", event);
