+++ src/entities/events/channel/mod.rs
@@ -1,0 +1,8 @@
+pub use {ack::*, delete::*, group_join::*, group_leave::*, start_typing::*, stop_typing::*};
+
+mod ack;
+mod delete;
+mod group_join;
+mod group_leave;
+mod start_typing;
+mod stop_typing;
