+++ src/entities/events/mod.rs
@@ -1,14 +1,6 @@
 pub(crate) use server_to_client::*;
-pub use {
-    channel_ack::*, channel_delete::*, channel_group_join::*, channel_group_leave::*,
-    channel_start_typing::*, channel_stop_typing::*, client_to_server::*,
-};
+pub use {channel::*, client_to_server::*};
 
-mod channel_ack;
-mod channel_delete;
-mod channel_group_join;
-mod channel_group_leave;
-mod channel_start_typing;
-mod channel_stop_typing;
+mod channel;
 mod client_to_server;
 mod server_to_client;
